+++ js/main.js
@@ -1,10 +1,11 @@
 import './draw-thumbnails.js';
 import './form.js';
-import { setDefaultSize } from './scale-operations.js';
+import { setImageDefaultSize } from './scale-operations.js';
 import { getData } from './api.js';
 import { createThumbnails } from './draw-thumbnails.js';
+import './effects.js';
 
-setDefaultSize();
+setImageDefaultSize();
 
 getData((pictures) => {
   createThumbnails(pictures);
